Initial functional version (0.1). Currently only has tested support for changing the solid fill color of an SVG node (gradients should work as well, but haven't been tested yet).
Decided to include a self-contained version of the QtSvg library because I had to make a few minor modifications to source there. So you should be able to compile and use this plugin without having to make any changes to Qt itself. I'll probably revisit this decision if I can get some feedback from the Qt developer community regarding how to manage a project like this. Who knows, could be worth proposing as a new feature for Qt 5.4. Or it could be that this is a horrible idea that leads to all sorts of performance woes and should be abandoned entirely...we'll see.

API is not even close to stable, just putting this up to get some eyes on it for now to get some feedback/help/advice, etc.